Made intake command and scheduled the new command
