+++ src/main/java/frc/robot/commands/HumanDrive.java
@@ -17,10 +17,10 @@ import frc.robot.subsystems.DriveConstants;
 // import frc.robot.CustomRotationHelper;
 import frc.robot.subsystems.DriveSubsystem;
 import frc.robot.subsystems.VisionSubsystem;
-import frc.robot.subsystems.DriveSubsystem.DriveState;
+// import frc.robot.subsystems.DriveSubsystem.DriveState;
 
 import com.playingwithfusion.TimeOfFlight;
-import com.playingwithfusion.TimeOfFlight.Status;
+// import com.playingwithfusion.TimeOfFlight.Status;
 
 public class HumanDrive extends Command {
